CropByCentroid/requirements.txt
ultralytics
opencv-python
pillow
numpy
scipy
tqdm